+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * Add `SnakeCaseToCamelCaseNameConverter`
  * Support subclasses of `\DateTime` and `\DateTimeImmutable` for denormalization
  * Support serialized names and paths configuration per group
+ * Add the `UidNormalizer::NORMALIZATION_FORMAT_RFC9562` constant
 
 7.1
 ---